requirements.txt
cached-property==1.5.2
et-xmlfile==1.1.0
lxml==5.3.0
openpyxl==3.1.5
packaging==24.1
pillow==10.4.0
PyPDF2==3.0.1
pytesseract==0.3.13
python-docx==1.1.2
python-dotenv==1.0.1
six==1.16.0
typing_extensions==4.12.2
Whoosh-Reloaded==2.7.5
langdetect==1.0.9
beautifulsoup4==4.12.3
EbookLib==0.18.0
mobi==0.3.3
colorlog==6.8.2
pydub==0.25.1
SpeechRecognition==3.10.4